add head() and tail() subs
head( $n, @list ) grabs the first $n elements from @list, or if $n is
negative, all but the last $n elements.

tail( $n, @list ) grabs the last $n elements from @list, or if $n is
negative, all but the first $n elements.

These are useful if you need to slice a list and do not want to assign
it to a temporary array. For example, as part of a map/grep/sort/etc...
chain.