List::Util::pairmap: extend stack
If pairmap generated more elements than it consumed, the stack could
overrun.

Because MULTICALLs are run on a separate stack, accumulate the results in
a temporary AV and push its contents onto the outer stack only at the end.

Note that the XS-sub branch of this function wasn't vulnerable, as
without MULTICALL, a single stack is used, and so will have have already
been extended by the XS returning its results.