increase accuracy of uniqnum
Using sprintf "%.15g" as the value to compare does not maintain the full
accuracy of floating point numbers.  Rather than 15, use a precision
value based on the accuracy the current platform can represent.

Also add some special cases: treat 0 and -0.0 as equivalent, and NaNs
are compared based on the platform's stringification of them.

Additionally, sprintf may not be able to accurately represent some
floating point numbers, no matter how many digits are given.  If the
numbers are outside the IV/UV range, they don't need to be compared to
anything other than other floats, so we can use a totally different
format for the unique string.  Use their raw bytes, as this is the
easiest and fastest.