Scalar-List-Utils 1.54 breaks on Win32 because trunc() is not available
not sure why trunc() is really needed, but if you are going to
use it you need to

   #include <math.h>

at least on Win32 builds. This caused havoc with all Pull Requests
to Perl/perl5. See https://github.com/Perl/perl5/issues/17550.

I reverted the push of 1.54 into core. This needs to be fixed
before it is updated.