always allow isvstring to be used
On perl 5.6, the isvstring function would not allow itself to be called
or exported. This behavior is annoying to implement, and forces loading
Exporter::Heavy in all cases.

Instead, isvstring can just always return false on perl 5.6. This is an
accurate behavior, as nothing on perl 5.6 is a vstring. And the odds of
anyone caring about this specific behavior on a prehistoric version of
perl is almost certainly 0.